backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-docx
openai
pydantic